block_ra harmonization
Conditions function does not work as expected. No condition names are assigned when the conditions function is used. The code to assign conditions as exists in simple_ra, complete_ra, etc. is fairly obviously missing, so this is the certainly the issue.

Some errors currently allowed to pass through the functions with regard to condition names, use of multiple conflicting options, etc.. These have been harmonized with the R version where possible.